assistant_demo/pyproject.toml
[tool.poetry]
name = "assistant_demo"
version = "0.1.0"
description = ""
authors = ["Konstantin Krauss <konsti@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
unique-sdk = "^0.7.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
